backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
asyncpg==0.29.0
redis[hiredis]==5.1.1
python-dotenv==1.0.1
email-validator==2.2.0
